Correctly initialize lastCommittedOffset.
The default initialization of lastCommittedOffset to 0 is a valid
offset.

This means that if the consumer only consumes the first ever message
on a topic before quitting, the offset will not be committed and the
next time the consumer starts it will start from scratch again.

That's what I get for writing trivial "send one message and see what
happens" tests.